moved zk initialization code from KubernetesLauncher to JobMaster

+++ twister2/resource-scheduler/src/java/edu/iu/dsc/tws/rsched/schedulers/k8s/KubernetesLauncher.java
@@ -27,11 +27,7 @@ import edu.iu.dsc.tws.api.exceptions.Twister2Exception;
 import edu.iu.dsc.tws.api.scheduler.ILauncher;
 import edu.iu.dsc.tws.api.scheduler.SchedulerContext;
 import edu.iu.dsc.tws.api.scheduler.Twister2JobState;
-import edu.iu.dsc.tws.common.zk.ZKBarrierManager;
 import edu.iu.dsc.tws.common.zk.ZKContext;
-import edu.iu.dsc.tws.common.zk.ZKEphemStateManager;
-import edu.iu.dsc.tws.common.zk.ZKEventsManager;
-import edu.iu.dsc.tws.common.zk.ZKPersStateManager;
 import edu.iu.dsc.tws.common.zk.ZKUtils;
 import edu.iu.dsc.tws.master.IJobTerminator;
 import edu.iu.dsc.tws.master.JobMasterContext;
@@ -102,15 +98,6 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
       return state;
     }
 
-    // create znodes at ZooKeeper server if ZK server is used
-    if (ZKContext.isZooKeeperServerUsed(config)) {
-      boolean jobZnodeCreated = createJobZnodes(job);
-      if (!jobZnodeCreated) {
-        // nothing to clear at this point, if the job znode is not created
-        return state;
-      }
-    }
-
     RequestObjectBuilder.init(config, job.getJobId(), jobFileSize);
     JobMasterRequestObject.init(config, job.getJobId(), jobFileSize);
 
@@ -151,31 +138,6 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
     state.setRequestGranted(true);
     return state;
   }
-
-  /**
-   * create znodes on ZooKeeper server for this job
-   */
-  public boolean createJobZnodes(JobAPI.Job job) {
-
-    CuratorFramework client = ZKUtils.connectToServer(ZKContext.serverAddresses(config));
-    String rootPath = ZKContext.rootNode(config);
-
-    try {
-      ZKEphemStateManager.createEphemDir(client, rootPath, job.getJobId());
-      ZKPersStateManager.createPersStateDir(client, rootPath, job);
-      ZKEventsManager.createEventsZNode(client, rootPath, job.getJobId());
-      ZKBarrierManager.createBarrierDir(client, rootPath, job.getJobId());
-
-      jobSubmissionStatus.setJobZNodesCreated(true);
-      return true;
-
-    } catch (Exception e) {
-      LOG.severe(e.getMessage()
-          + "\n++++++++++++++++++ Aborting submission ++++++++++++++++++");
-      return false;
-    }
-  }
-
 
   private String getJobMasterIP(JobAPI.Job job) {
     if (JobMasterContext.jobMasterRunsInClient(config)) {
@@ -292,9 +254,10 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
       serviceNames.add(jobMasterServiceName);
     }
 
-    boolean serviceExists = controller.existServices(serviceNames);
-    if (serviceExists) {
-      LOG.severe("Another job might be running. "
+    String existingService = controller.existServices(serviceNames);
+    if (existingService != null) {
+      LOG.severe("There is already a service with the name: " + existingService
+          + "\nAnother job might be running. "
           + "\nFirst terminate that job or create a job with a different name."
           + "\n++++++++++++++++++ Aborting submission ++++++++++++++++++");
       return false;
@@ -341,22 +304,6 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
       if (!secretExists) {
         LOG.severe("No Secret object is available in the cluster with the name: " + secretName
             + "\nFirst create this object or make that object created by your cluster admin."
-            + "\n++++++++++++++++++ Aborting submission ++++++++++++++++++");
-        return false;
-      }
-    }
-
-    // when fault tolerance is used, we check whether the znodes that will be created for this job
-    // is available at ZooKeeper server
-    if (ZKContext.isZooKeeperServerUsed(config)) {
-      CuratorFramework zkClient = ZKUtils.connectToServer(ZKContext.serverAddresses(config));
-      String rootPath = ZKContext.rootNode(config);
-      boolean jobZNodesExist = ZKUtils.isThereJobZNodes(zkClient, rootPath, jobID);
-
-      if (jobZNodesExist) {
-        LOG.severe("Some znodes already exist in ZooKeeper that will be used for this job. "
-            + "\nFirst terminate the previously running job with the same name or "
-            + "submit the job with a different job name. "
             + "\n++++++++++++++++++ Aborting submission ++++++++++++++++++");
         return false;
       }
@@ -630,10 +577,6 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
       String pvcName = KubernetesUtils.createPersistentVolumeClaimName(jobID);
       boolean claimDeleted = controller.deletePersistentVolumeClaim(pvcName);
     }
-
-    if (jobSubmissionStatus.isJobZNodesCreated()) {
-      deleteJobZnode(jobID);
-    }
   }
 
   /**
@@ -666,7 +609,13 @@ public class KubernetesLauncher implements ILauncher, IJobTerminator {
 
     // delete job znode from ZooWorker server if fault tolerant
     if (ZKContext.isZooKeeperServerUsed(config)) {
-      deleteJobZnode(jobID);
+      // if zk is running as a service,
+      // do not try to delete job znodes from zk
+      String serverAddresses = ZKContext.serverAddresses(config);
+      String sName = serverAddresses.substring(0, serverAddresses.indexOf("."));
+      if (controller.existServices(Arrays.asList(sName)) == null) {
+        deleteJobZnode(jobID);
+      }
     }
 
     return true;
